requirements.txt
fastmcp
uvicorn
pydantic
PyYAML
httpx
pdfplumber
jsonschema
python-multipart
jsonschema
pytesseract
Pillow
pytest
google-genai
python-docx
anthropic
